requirements.txt
allure-python-commons~=2.13.5
selenium~=4.29.0
pytest~=8.3.4